// components/DynamicIslandBanner/interfaces/notification-gooey.interface.ts
// Vendored from rit3zh/expo-dynamic-notifications @ 5de059a (MIT License). Imports rewritten from @/ alias to relative paths.
interface IBuildGooMatrix {
  gain: number;
  threshold: number;
}

interface INotificationGooey {
  blur?: number;
  gain?: number;
  threshold?: number;
  islandColor?: string;
  cardColor?: string;
  shadowColor?: string;
}

export type { IBuildGooMatrix, INotificationGooey };